Unit 1: useMark + Plot rewrite (new contract; legacy stack preserved)
Introduce the new <Plot>/useMark contract from the faithful-port plan:
imperative `plot()` does all rendering, React marks become 5-line factory
façades. The legacy stack is renamed to *.legacy.ts and kept intact so
existing marks continue to pass while subsequent units migrate them off.

- Rename old Plot.tsx, useMark.ts, PlotContext.ts to *.legacy.*
- Update all mark/legend/interaction imports to point at the legacy files
- Add new slim Plot.tsx, useMark.ts, PlotContext.ts (Option B)
- index.tsx exports legacy Plot as default and new Plot as PlotV2
- Add __validate.tsx scaffolding plus a JSDOM-backed test that renders
  Frame through the new contract and asserts a single <svg> with a
  black-stroked <rect>
