+++ src/react/__validate.tsx
@@ -1,0 +1,24 @@
+// Throwaway validation scaffolding for the new <Plot> + useMark contract.
+// Renders a single Frame mark via the new façade so the unit test can confirm
+// the imperative mount path works end-to-end. Delete once real marks have
+// migrated off the legacy stack.
+import React from "react";
+import {Plot} from "./Plot.js";
+import {useMark, stampOptions} from "./useMark.js";
+import {frame as frameMark} from "../marks/frame.js";
+
+function FrameNew(props: Record<string, any>) {
+  useMark({
+    stamp: stampOptions("frame", null, props),
+    factory: () => frameMark(props)
+  });
+  return null;
+}
+
+export function validatePlot() {
+  return (
+    <Plot width={200} height={100}>
+      <FrameNew stroke="black" />
+    </Plot>
+  );
+}
