+++ src/react/index.tsx
@@ -14,11 +14,15 @@
 //   }
 
 // Root component
-export {Plot} from "./Plot.js";
-export type {PlotProps} from "./Plot.js";
+// NOTE: During the faithful-port migration, the legacy Plot is still the
+// default export so existing marks (which use the legacy stack) continue to
+// work. The new contract is available as PlotV2 alongside the new useMark.
+export {Plot} from "./Plot.legacy.js";
+export type {PlotProps} from "./Plot.legacy.js";
+export {Plot as PlotV2} from "./Plot.js";
 
 // Context (for advanced usage / custom marks)
-export {PlotContext, FacetContext, usePlotContext, useFacetContext} from "./PlotContext.js";
+export {PlotContext, FacetContext, usePlotContext, useFacetContext} from "./PlotContext.legacy.js";
 export type {
   PlotContextValue,
   FacetContextValue,
@@ -27,11 +31,11 @@ export type {
   Dimensions,
   FacetInfo,
   ChannelSpec
-} from "./PlotContext.js";
+} from "./PlotContext.legacy.js";
 
 // Core hook (for building custom marks)
-export {useMark} from "./useMark.js";
-export type {UseMarkOptions, UseMarkResult} from "./useMark.js";
+export {useMark} from "./useMark.legacy.js";
+export type {UseMarkOptions, UseMarkResult} from "./useMark.legacy.js";
 
 // Style utilities (for building custom marks)
 export {
