+++ src/react/marks/Contour.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import {contours as d3Contours, geoPath} from "d3";
-import {useMark} from "../useMark.js";
+import {useMark} from "../useMark.legacy.js";
 import {indirectStyleProps, directStyleProps, isColorChannel, isColorValue} from "../styles.js";
-import type {ChannelSpec} from "../PlotContext.js";
+import type {ChannelSpec} from "../PlotContext.legacy.js";
 
 const defaults = {
   ariaLabel: "contour",
