+++ src/react/marks/Image.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import {useMark} from "../useMark.js";
+import {useMark} from "../useMark.legacy.js";
 import {
   indirectStyleProps,
   directStyleProps,
@@ -7,7 +7,7 @@ import {
   computeTransform,
   computeFrameAnchor
 } from "../styles.js";
-import type {ChannelSpec} from "../PlotContext.js";
+import type {ChannelSpec} from "../PlotContext.legacy.js";
 
 const defaults = {
   ariaLabel: "image",
