+++ src/react/marks/Rect.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import {useMark} from "../useMark.js";
+import {useMark} from "../useMark.legacy.js";
 import {indirectStyleProps, directStyleProps, channelStyleProps, isColorChannel, isColorValue} from "../styles.js";
 import {maybeTrivialIntervalX, maybeTrivialIntervalY} from "../../core/index.js";
-import type {ChannelSpec} from "../PlotContext.js";
+import type {ChannelSpec} from "../PlotContext.legacy.js";
 
 const defaults = {
   ariaLabel: "rect"
